collect tools descriptions in the env and pass it to the node prompts
