+++ conf/agent/plan_react_with_reasoner.yaml
@@ -11,6 +11,8 @@ templates:
     Keep your replies concise and direct. Prioritize clarity and avoid over-elaboration.
     Do not express emotions or opinions about user questions.
   allowed_steps: |
+    You have access to the following tools:
+    {tools_description}
     You are allowed to produce ONLY steps with the following JSON schemas:
     {allowed_steps}
     Do not reproduce the schema when producing steps; use it as a reference.
