+++ conf/agent/web_plan_react.yaml
@@ -10,6 +10,8 @@ templates:
     Keep your replies concise and direct. Prioritize clarity and avoid over-elaboration.
     Do not express emotions or opinions about user questions.
   allowed_steps: |
+    You have access to the following tools:
+    {tools_description}
     You are allowed to produce ONLY steps with the following JSON schemas:
     {allowed_steps}
     Do not reproduce the schema when producing steps; use it as a reference.
