+++ examples/gaia_agent/environment.py
@@ -4,7 +4,6 @@ from tapeagents.environment import ToolCollectionEnvironment
 from tapeagents.tools.base import Multitool, Tool
 from tapeagents.tools.browser import Browser
 from tapeagents.tools.code_executor import CodeExecutor
-from tapeagents.tools.computer.remote import RemoteComputer
 from tapeagents.tools.media_reader import VideoReader
 from tapeagents.tools.simple_browser import SimpleBrowser
 from tapeagents.tools.web_search import WebSearch
@@ -38,16 +37,5 @@ def get_env(
             CodeExecutor(exp_path=exp_path),
             VideoReader(exp_path=exp_path),
             SimpleBrowser(exp_path=exp_path, kwargs=kwargs) if simple_browser else Browser(exp_path=exp_path, **kwargs),
-        ]
-    )
-
-
-def get_computer_env(exp_path: str, **kwargs) -> ToolCollectionEnvironment:
-    return ToolCollectionEnvironment(
-        tools=[
-            WebSearch(),
-            CodeExecutor(exp_path=exp_path),
-            VideoReader(exp_path=exp_path),
-            RemoteComputer(exp_path=exp_path, **kwargs),
         ]
     )
