+++ tapeagents/agent.py
@@ -202,6 +202,7 @@ class Agent(BaseModel, Generic[TapeType]):
         description="List of nodes in the agent, order of the list used to determine the priority during activation. Nodes must have unique names.",
     )
     known_actions: list[type[Action]] = Field(default_factory=list)
+    tools_description: str = ""
     max_iterations: int = 100
     store_llm_calls: bool = False
