+++ tapeagents/tools/computer/computer.py
@@ -1,74 +1,77 @@
+import atexit
+import base64
 import logging
-from typing import Any, Literal
+import os
+import time
 
-from pydantic import Field
+import podman
+import requests
+from PIL import Image
 
-from tapeagents.core import Action, Observation
+from tapeagents.core import Action
 from tapeagents.tools.base import Multitool
-from tapeagents.tools.computer.api import (
-    get_cursor_position,
-    key_press,
-    mouse_click,
-    mouse_drag,
-    mouse_move,
-    open_url,
-    run_command,
-    type_text,
-)
 from tapeagents.tools.computer.steps import (
     ComputerObservation,
     GetCursorPositionAction,
     KeyPressAction,
     MouseClickAction,
-    MouseDragAction,
+    MouseClickAtAction,
+    MouseHoverAction,
     MouseMoveAction,
     OpenUrlAction,
+    PageDownAction,
+    PageUpAction,
     RunTerminalCommand,
     TypeTextAction,
 )
+from tapeagents.tools.grounding import GroundingModel
 
 logger = logging.getLogger(__name__)
-
-MAX_SCALING_TARGETS = {
-    "XGA": (1024, 768),  # 4:3
-    "WXGA": (1280, 800),  # 16:10
-    "FWXGA": (1366, 768),  # ~16:9
-}
-
-
-def chunks(s: str, chunk_size: int) -> list[str]:
-    return [s[i : i + chunk_size] for i in range(0, len(s), chunk_size)]
+logger.setLevel(logging.INFO)
+formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(name)s:%(lineno)d - %(funcName)s - %(message)s")
+console_handler = logging.StreamHandler()
+console_handler.setFormatter(formatter)
+logger.addHandler(console_handler)
 
 
 class Computer(Multitool):
-    actions: tuple[type[Action], ...] = (
-        KeyPressAction,
-        TypeTextAction,
-        MouseMoveAction,
-        MouseClickAction,
-        MouseDragAction,
-        GetCursorPositionAction,
-        OpenUrlAction,
-        RunTerminalCommand,
-    )
-    observations: tuple[type[Observation], ...] = (ComputerObservation,)
+    """
+    Computer tool for executing actions in the container with linux desktop.
+    Can execute linux terminal commands, open URLs in the browser.
+    Can move mouse, click, scroll, type text, enter hotkeys.
+    Can also take screenshots and get cursor position.
+    """
 
-    width: int = Field(description="Screen width", default=MAX_SCALING_TARGETS["XGA"][0])
-    height: int = Field(description="Screen height", default=MAX_SCALING_TARGETS["XGA"][1])
-    scaling_enabled: bool = Field(default=True, description="Enable resolution scaling")
-    typing_delay_ms: int = 180
+    exp_path: str | None = None
+    actions: tuple[type[Action], ...] = ()
+    observations: tuple[type[ComputerObservation], ...] = (ComputerObservation,)
+    computer_url: str
+    use_grounding: bool = True
+    grounding_api_url: str
+    container_image: str = "computer"
+    container_name: str = "computer"
 
-    def model_post_init(self, __context: Any) -> None:
+    def model_post_init(self, __context):
+        self._grounding = GroundingModel(url=self.grounding_api_url)
+        self._screenshot_dir = f"{self.exp_path}/attachments/remote_screenshots/"
+        os.makedirs(self._screenshot_dir, exist_ok=True)
         self._action_map = {
-            KeyPressAction: self._handle_key_press,
-            TypeTextAction: self._handle_type_text,
-            MouseMoveAction: self._handle_mouse_move,
-            MouseClickAction: self._handle_mouse_click,
-            MouseDragAction: self._handle_mouse_drag,
-            GetCursorPositionAction: self._handle_get_cursor_position,
-            OpenUrlAction: self._handle_open_url,
-            RunTerminalCommand: self._handle_command,
+            TypeTextAction: self.remote_execute_action,
+            OpenUrlAction: self.remote_execute_action,
+            KeyPressAction: self.remote_execute_action,
+            PageUpAction: self.page_up,
+            PageDownAction: self.page_down,
+            RunTerminalCommand: self.remote_execute_action,
         }
+        if self.use_grounding:
+            self._action_map[MouseClickAtAction] = self.mouse_click
+            self._action_map[MouseHoverAction] = self.mouse_hover
+        else:
+            self._action_map[MouseClickAction] = self.remote_execute_action
+            self._action_map[MouseMoveAction] = self.remote_execute_action
+            self._action_map[GetCursorPositionAction] = self.remote_execute_action
+        self.actions = tuple(self._action_map.keys())
+        launch_container(self.container_image, self.container_name, stop_at_exit=True)
 
     def execute_action(self, action: Action) -> ComputerObservation:
         action_type = type(action)
@@ -76,93 +79,100 @@ class Computer(Multitool):
             return self._action_map[action_type](action)
         raise ValueError(f"Unknown action type: {action_type}")
 
-    def _handle_command(self, action: RunTerminalCommand) -> ComputerObservation:
+    def mouse_hover(self, action: MouseHoverAction) -> ComputerObservation:
+        x, y = self._grounding.get_coords(self.get_screen(), action.element_description)
+        return self.remote_execute_action(MouseMoveAction(x=int(x), y=int(y)))
+
+    def mouse_click(self, action: MouseClickAtAction) -> ComputerObservation:
+        x, y = self._grounding.get_coords(self.get_screen(), action.element_description)
+        return self.remote_execute_action(MouseClickAction(x=int(x), y=int(y), button=action.button))
+
+    def page_up(self, action: PageUpAction) -> ComputerObservation:
+        return self.remote_execute_action(KeyPressAction(text="Page_Up"))
+
+    def page_down(self, action: PageDownAction) -> ComputerObservation:
+        return self.remote_execute_action(KeyPressAction(text="Page_Down"))
+
+    def remote_execute_action(self, action: Action) -> ComputerObservation:
+        payload = {"kind": action.kind, "params": action.model_dump()}
         try:
-            obs_dict = run_command(action.command, wait_output=action.wait_output)
-            obs = ComputerObservation(**obs_dict)
-        except Exception as e:
-            obs = ComputerObservation(error=f"Command {action.command} failed: {e}")
+            response = requests.post(f"{self.computer_url}/execute", json=payload)
+            response.raise_for_status()
+            obs_dict = response.json()
+            logger.info(f"Received observation: {obs_dict.keys()}")
+            return self.save_screenshot(ComputerObservation(**obs_dict))
+        except requests.exceptions.RequestException as e:
+            logger.error(f"API request failed: {str(e)}")
+            return ComputerObservation(image_path="", error=f"API request failed: {str(e)}")
+
+    def save_screenshot(self, obs: ComputerObservation) -> ComputerObservation:
+        if obs.base64_image:
+            fname = f"{self._screenshot_dir}/screen_{int(time.time())}.png"
+            with open(fname, "wb") as f:
+                f.write(base64.b64decode(obs.base64_image))
+            obs.image_path = fname
+            obs.base64_image = None
         return obs
 
-    def _handle_open_url(self, action: OpenUrlAction) -> ComputerObservation:
-        try:
-            obs_dict = open_url(action.url)
-            obs = ComputerObservation(**obs_dict)
-        except Exception as e:
-            obs = ComputerObservation(error=f"Open URL {action.url} failed: {e}")
-        return obs
+    def get_screen(self) -> Image:
+        obs = self.remote_execute_action(GetCursorPositionAction())
+        if obs.error:
+            raise ValueError(f"Failed to get screen: {obs.error}")
+        return Image.open(obs.image_path)
 
-    def _handle_key_press(self, action: KeyPressAction) -> ComputerObservation:
-        try:
-            obs_dict = key_press(action.text)
-            obs = ComputerObservation(**obs_dict)
-        except Exception as e:
-            obs = ComputerObservation(error=f"Key press {action.text} failed: {e}")
-        return obs
+    def reset(self):
+        self.remote_execute_action(RunTerminalCommand(command='xdotool search "" windowkill %@'))
 
-    def _handle_type_text(self, action: TypeTextAction) -> ComputerObservation:
-        try:
-            obs_dict = type_text(action.text, delay_ms=self.typing_delay_ms)
-            obs = ComputerObservation(**obs_dict)
-        except Exception as e:
-            obs = ComputerObservation(error=f"Typing failed: {e}")
-        return obs
+    def close(self):
+        return super().close()
 
-    def _handle_mouse_move(self, action: MouseMoveAction) -> ComputerObservation:
-        x, y = self._scale_coordinates("api", action.x, action.y)
-        try:
-            obs_dict = mouse_move(x, y)
-            obs = ComputerObservation(**obs_dict)
-        except Exception as e:
-            obs = ComputerObservation(error=f"Mouse move failed: {e}")
-        return obs
 
-    def _handle_mouse_click(self, action: MouseClickAction) -> ComputerObservation:
-        try:
-            obs_dict = mouse_click(action.x, action.y, button=action.button)
-            obs = ComputerObservation(**obs_dict)
-        except Exception as e:
-            obs = ComputerObservation(error=f"Mouse click failed: {e}")
-        return obs
+def launch_container(container_image="computer", container_name="computer", stop_at_exit=False):
+    podman_client = podman.from_env()
+    try:
+        container = podman_client.containers.get(container_name)
+    except podman.errors.NotFound:
+        logger.info(f"Creating container from image '{container_image}'")
+        home_dir = os.path.join(os.getcwd(), ".computer")
+        zip_home = os.path.join(os.path.dirname(__file__), "home.zip")
+        if os.path.exists(home_dir):
+            os.system(f"rm -rf {home_dir}")
+            logger.info(f"Removed existing home directory: {home_dir}")
+        else:
+            os.makedirs(home_dir)
+        os.system(f"unzip -qq -o {zip_home} -d {home_dir}")
+        logger.info(f"Recreated home directory from zip: {zip_home}")
+        container = podman_client.containers.create(
+            container_image,
+            name=container_name,
+            auto_remove=True,
+            ports={
+                "3000": 3000,
+                "8000": 8000,
+            },
+            mounts=[
+                {
+                    "type": "bind",
+                    "source": os.path.join(home_dir, "home"),
+                    "target": "/config",
+                }
+            ],
+        )
+        container.start()
+        logger.info("Starting..")
+    while container.status != "running":
+        time.sleep(0.2)
+        container.reload()
+    logger.info("Container ready")
 
-    def _handle_mouse_drag(self, action: MouseDragAction) -> ComputerObservation:
-        x, y = self._scale_coordinates("api", action.x, action.y)
-        try:
-            obs_dict = mouse_drag(x, y)
-            obs = ComputerObservation(**obs_dict)
-        except Exception as e:
-            obs = ComputerObservation(error=f"Mouse drag failed: {e}")
-        return obs
+    if stop_at_exit:
 
-    def _handle_get_cursor_position(self, action: GetCursorPositionAction) -> ComputerObservation:
-        try:
-            obs_dict = get_cursor_position()
-            obs = ComputerObservation(**obs_dict)
-        except Exception as e:
-            obs = ComputerObservation(error=f"Get cursor position failed: {e}")
-        return obs
+        def cleanup() -> None:
+            try:
+                logger.info(f"Stopping container {container.name}")
+                container.stop()
+            except podman.errors.NotFound:
+                pass
+            atexit.unregister(cleanup)
 
-    def _scale_coordinates(self, source: Literal["api", "computer"], x: int, y: int) -> tuple[int, int]:
-        """Scale coordinates based on target resolution"""
-        if not self.scaling_enabled:
-            return x, y
-        ratio = self.width / self.height
-        target_dimension = None
-        for width, height in MAX_SCALING_TARGETS.values():
-            # allow some error in the aspect ratio - not ratios are exactly 16:9
-            if abs(width / height - ratio) < 0.02:
-                if width < self.width:
-                    target_dimension = (width, height)
-                break
-        if target_dimension is None:
-            return x, y
-        # should be less than 1
-        x_scaling_factor = target_dimension[0] / self.width
-        y_scaling_factor = target_dimension[1] / self.height
-        if source == "api":
-            if x > self.width or y > self.height:
-                raise Exception(f"Coordinates {x}, {y} are out of bounds")
-            # scale up
-            return round(x / x_scaling_factor), round(y / y_scaling_factor)
-        # scale down
-        return round(x * x_scaling_factor), round(y * y_scaling_factor)
+        atexit.register(cleanup)
